+++ SharpSource/SharpSource.Test/Helpers/CSCodeFix.cs
@@ -27,12 +27,21 @@ public static partial class CSharpCodeFixVerifier<TAnalyzer, TCodeFix>
         => CSharpCodeFixVerifier<TAnalyzer, TCodeFix, MSTestVerifier>.Diagnostic(descriptor).WithLocation(location);
 
     /// <inheritdoc cref="CodeFixVerifier{TAnalyzer, TCodeFix, TTest, TVerifier}.VerifyAnalyzerAsync(string, DiagnosticResult[])"/>
-    public static async Task VerifyAnalyzerAsync(string source, params DiagnosticResult[] expected)
+    public static async Task VerifyAnalyzerAsync(string source, string[]? additionalFiles = null, params DiagnosticResult[] expected)
     {
         var test = new Test
         {
             TestCode = source,
         };
+
+        if (additionalFiles != null)
+        {
+            foreach (var file in additionalFiles)
+            {
+                var filename = Guid.NewGuid().ToString();
+                test.TestState.Sources.Add(($"{filename}.cs", file));
+            }
+        }
 
         test.ExpectedDiagnostics.AddRange(expected);
         await test.RunAsync(CancellationToken.None);